Move chain utils to own folder. Rename ecc folder to encrypt.
