+++ snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/fioprotocol/fio-wallet-snap.git"
   },
   "source": {
-    "shasum": "AkG1PlAP6j2l73AOUo9tSyAYWc+i93j7Gqohy/5yNPI=",
+    "shasum": "9iWhyjnmHiCoRdas3lVInajq/z/r/3VBTbgqzY++Hcw=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
